requirements.txt
python-telegram-bot==21.6
google-generativeai==0.8.3
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml>=5.0
requests==2.32.3
python-dotenv==1.0.1
